README and library version update
The README.md was updated to correctly show the test file.

For MacOS users, there are some known issues with pyglet and its use of openGL (https://github.com/openai/gym/issues/2101). To correct it, the pyglet 1.5.11 version can be used. For what I can tell this version didn't bring any incompatibility issues and it is more recent than the one that was in use.